[Decomp][ObjectMgr] Cohesion split: extract locale/quest-relation/instance/spell/text loaders to ObjectMgr*.cpp
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Splits five cohesive loader groups out of ObjectMgr.cpp:
  - ObjectMgrInstanceData.cpp  (instance template + arena teams + groups)
  - ObjectMgrLocales.cpp       (locale string loaders/accessors)
  - ObjectMgrQuestRelations.cpp(creature/gameobject quest relations)
  - ObjectMgrSpellData.cpp     (spell-script + related loaders)
  - ObjectMgrText.cpp          (gossip/npc text loaders)

Pure code move, no behaviour change. CMake file(GLOB) picks the new TUs up
automatically. Verified: game.lib compiles and links clean (MSVC, Release).
(The source's git numstat shows a few "added" lines -- a diff-alignment
artifact from several getters sharing the identical `return &itr->second;`
body; the getters remain intact, as the clean build confirms.)

Not ported (absent in MangosOne TBC 2.4.3 -- later-expansion methods):
  ObjectMgr::GetWorldTemplate, ObjectMgr::LoadInstanceEncounters,
  ObjectMgr::LoadNPCSpellClickSpells

Ported-from: mangosthree/server@c8462f109703
